requirements.txt
streamlit==1.45.0
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
langchain-community
langgraph
langchainhub
langchain-groq
sqlalchemy
aiosqlite 
eralchemy
graphviz
SpeechRecognition
pyttsx3
pyperclip